+++ app/Http/Controllers/UserController.php
@@ -10,14 +10,17 @@ use App\Http\Resources\UserResources;
 use App\Models\User;
 use App\Repositories\UserRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
     protected $userRepository;
+    public $user;
 
     public function __construct(UserRepository $userRepository)
     {
         $this->userRepository = $userRepository;
+        $this->user = Auth::user();
     }
 
     /**
@@ -28,25 +31,29 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
-        $keyword    = $request->input('keyword') ?? '';
-        $perPage    = $request->input('per_page') ?? 10;
-        $listUser   = User::with('roles')
-            ->when($keyword <> '', function($q) use ($keyword) {
-                $q->where('email', 'like', "%{$keyword}%")
-                    ->orWhere('username', 'like', "%{$keyword}%")
-                    ->orWhereHas('roles', function ($q) use ($keyword) {
-                        $q->where('name', 'like', "%{$keyword}%");
-                    });
-            })
-            ->orderBy('id', 'desc');
+        if ($this->user->can('show users')):
+            $keyword    = $request->input('keyword') ?? '';
+            $perPage    = $request->input('per_page') ?? 10;
+            $listUser   = User::with('roles')
+                ->when($keyword <> '', function($q) use ($keyword) {
+                    $q->where('email', 'like', "%{$keyword}%")
+                        ->orWhere('username', 'like', "%{$keyword}%")
+                        ->orWhereHas('roles', function ($q) use ($keyword) {
+                            $q->where('name', 'like', "%{$keyword}%");
+                        });
+                })
+                ->orderBy('id', 'desc');
 
-        $listUser = $perPage === 'all' ? $listUser->get():$listUser->paginate($perPage);
+            $listUser = $perPage === 'all' ? $listUser->get():$listUser->paginate($perPage);
 
-        if ($listUser) {
-            return UserResources::collection($listUser);
-        }
+            if ($listUser) {
+                return UserResources::collection($listUser);
+            }
 
-        return $this->dataEmpty('User');
+            return $this->dataEmpty('User');
+        else:
+            return $this->unAuthorized();
+        endif;
     }
 
     /**
@@ -56,57 +63,73 @@ class UserController extends Controller
      */
     public function store(UserStoreRequest $request)
     {
-        $data = $request->except(['password_confirmation', 'role']);
+        if ($this->user->can('store user')):
+            $data = $request->except(['password_confirmation', 'role']);
 
-        $data['password'] = app('hash')->make($data['password']);
+            $data['password'] = app('hash')->make($data['password']);
 
-        $user = User::create($data);
+            $user = User::create($data);
 
-        $user->assignRole($request->role);
+            $user->assignRole($request->role);
 
-        return $user ? $this->storeTrue('user'):$this->storeFalse('user');
+            return $user ? $this->storeTrue('user'):$this->storeFalse('user');
+        else:
+            return $this->unAuthorized();
+        endif;
     }
 
     public function edit($id)
     {
-        $user = User::find($id);
-        if ($user) {
-            return new UserDetailResources($user);
-        }
+        if ($this->user->can('edit user')):
+            $user = User::find($id);
+            if ($user) {
+                return new UserDetailResources($user);
+            }
 
-        return $this->dataNotFound('user');
+            return $this->dataNotFound('user');
+        else:
+            return $this->unAuthorized();
+        endif;
     }
 
     public function update(UserUpdateRequest $request, $id)
     {
-        $user = User::find($id);
+        if ($this->user->can('update user')):
+            $user = User::find($id);
 
-        if ($user) {
-            $password   = $request->password;
+            if ($user) {
+                $password   = $request->password;
 
-            $data       = $request->except(['password', 'password_confirmation', 'id', 'role', 'created_at']);
+                $data       = $request->except(['password', 'password_confirmation', 'id', 'role', 'created_at']);
 
-            if ($password != '' && $password != null) {
-                $data['password'] = app('hash')->make($password);
+                if ($password != '' && $password != null) {
+                    $data['password'] = app('hash')->make($password);
+                }
+
+                $user->assignRole($request->role);
+
+                return $user->update($data) ? $this->updateTrue('user') : $this->updateFalse('user');
             }
 
-            $user->assignRole($request->role);
-
-            return $user->update($data) ? $this->updateTrue('user') : $this->updateFalse('user');
-        }
-
-        return $this->dataNotFound('user');
+            return $this->dataNotFound('user');
+        else:
+            return $this->unAuthorized();
+        endif;
     }
 
     public function destroy($id)
     {
-        $user = User::find($id);
+        if ($this->user->can('delete user')):
+            $user = User::find($id);
 
-        if ($user) {
-            return $user->delete() ? $this->destroyTrue('pengguna') : $this->destroyFalse('pengguna');
-        }
+            if ($user) {
+                return $user->delete() ? $this->destroyTrue('pengguna') : $this->destroyFalse('pengguna');
+            }
 
-        return $this->dataNotFound('user');
+            return $this->dataNotFound('user');
+        else:
+            return $this->unAuthorized();
+        endif;
     }
 
     public function me()
